po dodaniu zamówienia już się nie wylogowuje, dodane rozmiary butów z sortowaniem po ilości, drobne poprawki typu nie da sie dodać "pustego" buta

+++ Projekt/src/main/java/com/example/projekt/Model/Zamowienie.java
@@ -19,8 +19,8 @@ public class Zamowienie {
     @ManyToOne
     @JoinColumn(name = "but_id")
     private But but;
-    private int rozmiar;
-    public Zamowienie(Klient klient, But but, int rozmiar){
+    private Double rozmiar;
+    public Zamowienie(Klient klient, But but, Double rozmiar){
         this.klient = klient;
         this.but = but;
         this.rozmiar = rozmiar;
